refactor(e2e): apply reputation #[path] decls + extract admin_config (sub-phases 3+5/7)
Sub-phase 3 — reputation #[path] declarations applied to e2e.rs:
  mod v1_rt_r3_fixtures; → #[path = "e2e/reputation_rt_r3.rs"] (1053-line body removed)
  mod v1_rt_r4_fixtures; → #[path = "e2e/reputation_rt_r4.rs"] (314-line body removed)
  (Files were staged in prior commit; e2e.rs inline mods now removed)

Sub-phase 5 — admin_config block extracted:
  tests/e2e/admin_config.rs: admin_config_fixtures mod + 24 top-level tests (2502 lines)
  e2e.rs uses include!("e2e/admin_config.rs")

Compile gate: SUB5_COMPILE_EXIT_0 (cargo-test.bat --no-run).
e2e.rs: 12,044 → 8,188 lines.
